requirements.txt
# Email Monitor - Dependencias de producción
# Core dependencies
websockets>=12.0
PyJWT>=2.8.0
bcrypt>=4.1.0
python-dotenv>=1.0.0

# Development dependencies (opcional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Security dependencies  
cryptography>=41.0.0

# Opcional: Para SSL/TLS certificates
certifi>=2023.5.7